Criação da função creatStoreByCep como também sua route e seu dto, e atualização nas funções de findById e deleteStore, storeID agora é chave única e serve também paralelamente nas funções de findById e deleteStore, funcionando tanto com esse ID, como o do próprio banco de dados

+++ src/store/store.service.ts
@@ -1,9 +1,11 @@
 import { Injectable } from '@nestjs/common';
-import { Store, StoreSchema } from './model/store.schema';
-import { InjectModel, MongooseModule } from '@nestjs/mongoose';
+import { Store } from './model/store.schema';
+import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { CreateStoreDto } from './dto/createStoreDto';
 import { GoogleApiService } from 'src/utils/googleMapsService';
+import { CreateStoreByCepDto } from './dto/createStoreByCepDto';
+import { cepInfos } from 'src/utils/viaCepService';
 
 @Injectable()
 export class StoreService {
@@ -16,12 +18,17 @@ export class StoreService {
     }
 
     async findById(id: string): Promise<Store>{
+
+        if(!isNaN(Number(id))){
+            return await this.storeModel.findOne({storeID: id})
+        }    
+       
         return await this.storeModel.findById(id)
+        
     }
 
     async createStore(createStoreDto: CreateStoreDto): Promise<Store>{
         const newStore = new this.storeModel(createStoreDto)
-        newStore.country = 'Brasil'
         
         const coordenates = await GoogleApiService.getCordenates(createStoreDto.postalCode)
         newStore.latitude = coordenates.latitude
@@ -30,7 +37,29 @@ export class StoreService {
         return await newStore.save()
     }
 
+    async createStoreByCep(createStoreByCepDto: CreateStoreByCepDto): Promise<Store>{
+        const newStore = new this.storeModel(createStoreByCepDto)
+        const data = await cepInfos(createStoreByCepDto.postalCode)
+
+        newStore.address1 = data.logradouro
+        newStore.city = data.localidade
+        newStore.district = data.bairro
+        newStore.state = data.uf
+
+        const coordenates = await GoogleApiService.getCordenates(createStoreByCepDto.postalCode)
+
+        newStore.latitude = coordenates.latitude
+        newStore.longitude = coordenates.longitude  
+              
+        return await newStore.save()
+    }
+
     async deleteStore(id: string){
+
+        if(!isNaN(Number(id))){
+            return await this.storeModel.findOne({storeID: id})
+        }  
+
         return await this.storeModel.findByIdAndDelete(id)
     }
